fix: fixed comments from reviews
