+++ Frontend/src/Components/UI/Tile/Tile.tsx
@@ -1,5 +1,5 @@
 import { defineComponent } from "vue";
-import { cn } from "@/lib/utils";
+import { cn } from "@/lib/tailwind.utils";
 
 export default defineComponent({
   name: "Tile",
