+++ Frontend/src/lib/dates.utils.tsx
@@ -1,0 +1,37 @@
+type DateFormatType =
+  | "fullDateTime"
+  | "shortDate"
+  | "monthShort"
+  | "timeOnly"
+  | "defaultDate";
+
+export function formatDate(date: Date, format: DateFormatType): string {
+  switch (format) {
+    case "fullDateTime":
+      return date.toLocaleDateString("default", {
+        day: "numeric",
+        month: "long",
+        year: "numeric",
+        hour: "2-digit",
+        minute: "2-digit",
+        hour12: false,
+      });
+    case "shortDate":
+      return date.toLocaleDateString("default", {
+        day: "numeric",
+        month: "short",
+      });
+    case "monthShort":
+      return date.toLocaleDateString("default", {
+        month: "short",
+      });
+    case "timeOnly":
+      return date.toLocaleTimeString("default", {
+        hour: "2-digit",
+        minute: "2-digit",
+        hour12: false,
+      });
+    default:
+      return date.toLocaleDateString();
+  }
+}
