+++ Frontend/src/lib/images.tsx
@@ -1,0 +1,3 @@
+export default {
+  sticky_logo: "https://public.svsticky.nl/logos/hoofd_outline_wit.svg",
+};
